refactor: move factory utils to alternativeOrder state
