+++ apps/cowswap-frontend/src/modules/limitOrders/state/limitOrdersRawStateAtom.ts
@@ -1,4 +1,4 @@
-import { atom, Getter, PrimitiveAtom, SetStateAction, Setter, WritableAtom } from 'jotai'
+import { atom } from 'jotai'
 import { atomWithStorage } from 'jotai/utils'
 
 import { atomWithPartialUpdate } from '@cowprotocol/common-utils'
@@ -8,7 +8,7 @@ import { OrderKind, SupportedChainId } from '@cowprotocol/cow-sdk'
 import { DEFAULT_TRADE_DERIVED_STATE, TradeDerivedState } from 'modules/trade/types/TradeDerivedState'
 import { ExtendedTradeRawState, getDefaultTradeRawState } from 'modules/trade/types/TradeRawState'
 
-import { isAlternativeOrderModalVisibleAtom } from 'common/state/alternativeOrder'
+import { alternativeOrderAtomSetterFactory, alternativeOrderReadWriteAtomFactory } from 'common/state/alternativeOrder'
 
 export interface LimitOrdersDerivedState extends TradeDerivedState {
   readonly isUnlocked: boolean
@@ -59,51 +59,20 @@ const alternativeDerivedStateAtom = atom<LimitOrdersDerivedState>({
 
 // Pick atom according to type of form displayed
 
-export const limitOrdersRawStateAtom = readWriteAtomFactory<LimitOrdersRawState>(
+export const limitOrdersRawStateAtom = alternativeOrderReadWriteAtomFactory<LimitOrdersRawState>(
   regularRawStateAtom,
   alternativeRawStateAtom
 )
 
 export const updateLimitOrdersRawStateAtom = atom(
   null,
-  atomSetterFactory<
+  alternativeOrderAtomSetterFactory<
     null, // pass null to indicate there is no getter
     Partial<LimitOrdersRawState>
   >(regularUpdateRawStateAtom, alternativeUpdateRawStateAtom)
 )
 
-export const limitOrdersDerivedStateAtom = readWriteAtomFactory<LimitOrdersDerivedState>(
+export const limitOrdersDerivedStateAtom = alternativeOrderReadWriteAtomFactory<LimitOrdersDerivedState>(
   regularDerivedStateAtom,
   alternativeDerivedStateAtom
 )
-
-// utils
-
-function atomGetterFactory<AtomValue>(regular: PrimitiveAtom<AtomValue>, alternative: PrimitiveAtom<AtomValue>) {
-  return (get: Getter) => get(get(isAlternativeOrderModalVisibleAtom) ? alternative : regular)
-}
-
-type WritableWithOptionalSetterValue<GetterValue, SetterValue> = WritableAtom<GetterValue, [value: SetterValue], void>
-
-function atomSetterFactory<AtomGetterValue, AtomWriterParamValue>(
-  regular: WritableWithOptionalSetterValue<AtomGetterValue, AtomWriterParamValue>,
-  alternative: WritableWithOptionalSetterValue<AtomGetterValue, AtomWriterParamValue>
-) {
-  return (get: Getter, set: Setter, value: AtomWriterParamValue) => {
-    if (get(isAlternativeOrderModalVisibleAtom)) {
-      set(alternative, value)
-    } else {
-      set(regular, value)
-    }
-  }
-}
-
-function readWriteAtomFactory<AtomType>(
-  regular: WritableWithOptionalSetterValue<AtomType, SetStateAction<AtomType>>,
-  alternative: WritableWithOptionalSetterValue<AtomType, SetStateAction<AtomType>>
-) {
-  return atom(
-    atomGetterFactory<AtomType>(regular, alternative),
-    atomSetterFactory<AtomType, AtomType>(regular, alternative)
-  )
-}
